Replace attribute value derived from another attribute by a static value.
Chef does not correctly manage attribute for which the value is derived
from anoter value in the case of the derived value is overriden.

Therefore the following DN attributes are expressed relatively to the
basedn:
 * default.ca_openldap.rootdn
 * default.ca_openldap.ppolicy_default_config_dn

Recipes are updated in order to take into account this new semantic.
A new helper is defined in ca_openldap.rb to build the rootdn.