Fix error when creating DIT if TLS mode is not enable (#7)
Error was due to a wrong type comparison on the port (String vs Fixnum).
Additionnaly, determining if the TLS mode is enable based on the port
number is not correct.
A new method, tls_enabled? is added in libraries/ca_openldap to
determine if TLS mode is enable or not base on the
node.ca_opendldap.tls.enable attribute value.